Rename folder LoginComponent to login-component
